--- a/database/init/03-sample-data.cql
+++ b/database/init/03-sample-data.cql
@@ -1,4 +1,5 @@
 -- Insert sample data for testing
+USE beauty_salon;
 
 -- Sample services
 INSERT INTO services (id, name, description, duration, price, category, is_active, created_at, updated_at)
@@ -18,20 +19,20 @@ VALUES (uuid(), 'Limpeza de Pele', 'Limpeza facial profunda', 90, 80.00, 'Estét
 
 -- Sample staff
 INSERT INTO staff (id, name, email, phone, role, specialties, is_active, created_at, updated_at)
-VALUES (uuid(), 'Maria Silva', 'maria@salao.com', '(11) 99999-1111', 'Cabeleireira', {'Corte', 'Escova', 'Coloração'}, true, toTimestamp(now()), toTimestamp(now()));
+VALUES (uuid(), 'Maria Silva', 'maria@salao.com', '(11) 99999-1111', 'Cabeleireira', ['Corte', 'Escova', 'Coloração'], true, toTimestamp(now()), toTimestamp(now()));
 
 INSERT INTO staff (id, name, email, phone, role, specialties, is_active, created_at, updated_at)
-VALUES (uuid(), 'Ana Santos', 'ana@salao.com', '(11) 99999-2222', 'Manicure', {'Manicure', 'Pedicure', 'Nail Art'}, true, toTimestamp(now()), toTimestamp(now()));
+VALUES (uuid(), 'Ana Santos', 'ana@salao.com', '(11) 99999-2222', 'Manicure', ['Manicure', 'Pedicure', 'Nail Art'], true, toTimestamp(now()), toTimestamp(now()));
 
 INSERT INTO staff (id, name, email, phone, role, specialties, is_active, created_at, updated_at)
-VALUES (uuid(), 'Carla Oliveira', 'carla@salao.com', '(11) 99999-3333', 'Esteticista', {'Limpeza de Pele', 'Massagem', 'Depilação'}, true, toTimestamp(now()), toTimestamp(now()));
+VALUES (uuid(), 'Carla Oliveira', 'carla@salao.com', '(11) 99999-3333', 'Esteticista', ['Limpeza de Pele', 'Massagem', 'Depilação'], true, toTimestamp(now()), toTimestamp(now()));
 
--- Sample customers
+-- Sample customers (phone numbers in digits-only format for Python FastAPI compatibility)
 INSERT INTO customers (id, name, email, phone, address, created_at, updated_at)
-VALUES (uuid(), 'Julia Costa', 'julia@email.com', '(11) 88888-1111', 'Rua das Flores, 123', toTimestamp(now()), toTimestamp(now()));
-
-INSERT INTO customers (id, name, email, phone, address, created_at, updated_at)
-VALUES (uuid(), 'Fernanda Lima', 'fernanda@email.com', '(11) 88888-2222', 'Av. Paulista, 456', toTimestamp(now()), toTimestamp(now()));
+VALUES (uuid(), 'Julia Costa', 'julia@email.com', '11888881111', 'Rua das Flores, 123', toTimestamp(now()), toTimestamp(now()));
 
 INSERT INTO customers (id, name, email, phone, address, created_at, updated_at)
-VALUES (uuid(), 'Beatriz Rocha', 'beatriz@email.com', '(11) 88888-3333', 'Rua Augusta, 789', toTimestamp(now()), toTimestamp(now()));
+VALUES (uuid(), 'Fernanda Lima', 'fernanda@email.com', '11888882222', 'Av. Paulista, 456', toTimestamp(now()), toTimestamp(now()));
+
+INSERT INTO customers (id, name, email, phone, address, created_at, updated_at)
+VALUES (uuid(), 'Beatriz Rocha', 'beatriz@email.com', '11888883333', 'Rua Augusta, 789', toTimestamp(now()), toTimestamp(now()));