… DEPLOYMENT_GUIDE.md
… FINAL_STATUS_REPORT.md
… backend-python/tests/test_repositories.py
… backend-python/tests/test_repositories_async.py
… backend-python/tests/test_routers.py
… backend-python/tests/test_services.py
… frontend/coverage/clover.xml
… frontend/coverage/coverage-final.json
… frontend/coverage/lcov-report/index.html
… frontend/coverage/lcov-report/src/App.js.html
… frontend/coverage/lcov-report/src/components/Appointments.js.html
… frontend/coverage/lcov-report/src/components/Customers.js.html
… frontend/coverage/lcov-report/src/components/Dashboard.js.html
… frontend/coverage/lcov-report/src/components/Navbar.js.html
… frontend/coverage/lcov-report/src/components/Services.js.html
… frontend/coverage/lcov-report/src/components/Staff.js.html
… frontend/coverage/lcov-report/src/components/index.html
… frontend/coverage/lcov-report/src/index.html
… frontend/coverage/lcov-report/src/index.js.html
… frontend/coverage/lcov-report/src/services/api.js.html
… frontend/coverage/lcov-report/src/services/index.html
… frontend/coverage/lcov.info
… frontend/cypress/screenshots/advanced-flows.cy.js/Advanced E2E Flows -- Accessibility and Usability -- should be keyboard navigable (failed) (attempt 2).png
… frontend/cypress/screenshots/advanced-flows.cy.js/Advanced E2E Flows -- Accessibility and Usability -- should be keyboard navigable (failed) (attempt 3).png
… frontend/cypress/screenshots/advanced-flows.cy.js/Advanced E2E Flows -- Accessibility and Usability -- should be keyboard navigable (failed).png
… frontend/cypress/screenshots/advanced-flows.cy.js/Advanced E2E Flows -- Accessibility and Usability -- should have proper ARIA labels and roles (failed) (attempt 2).png
… frontend/cypress/screenshots/advanced-flows.cy.js/Advanced E2E Flows -- Accessibility and Usability -- should have proper ARIA labels and roles (failed) (attempt 3).png
… frontend/cypress/screenshots/advanced-flows.cy.js/Advanced E2E Flows -- Accessibility and Usability -- should have proper ARIA labels and roles (failed).png
… frontend/cypress/screenshots/advanced-flows.cy.js/Advanced E2E Flows -- Accessibility and Usability -- should work with screen reader simulation (failed) (attempt 2).png
… frontend/cypress/screenshots/advanced-flows.cy.js/Advanced E2E Flows -- Accessibility and Usability -- should work with screen reader simulation (failed) (attempt 3).png